Enable copy-to-clipboard for left-rome
Apparently because we are wrapping everything in a
`google-sheets-html-origin` element (which causes Sheets to parse the
data-sheets-formula attribute), we can no longer embed links in the
middle of td content and have it partially highlight the text. (In fact,
it no longer seems to automatically handle links in table cells at all).

So detect when a table cell has a top-level link and convert that to a
`=hyperlink` formula, and add copy-only versions of the DIYC
instructions in left-rome that have a link spanning the full cell.