Fix Paper Trail round page unlocks and tooltips
Incorporate a similar model to the Missing Diamond round to show unlock
modals for locked puzzles instead of links, and switch to using
floating-ui for tooltips.

In order to fix stacking of tooltips, I had to pull them outside of the
individual desk item button/link elements - those elements are using
`filter` for styling, which establishes a separate stacking context, so
if we want the tooltips to stack on top of everything, they need to live
outside of that context.