Add more features to hub page
This adds alt text to all of the assets already on the hub page. It also
adds the main question text (I don't think we will have anything to link
this too so we can punt on that) and the "AKA" taped text for each of
the four side investigations.

Previously drop shadows were added on all assets, but assets
representing new writing on existing paper shouldn't have drop shadows,
so there's an option to exclude that.

I also found that the default positioning of rotated objects was wrong,
because while CSS rotates from the center by default, Affinity Designer
rotates from the top-left. I went back and fixed the position of a few
assets to match the original mockups for ease of layering additional
assets on top of them, but mostly left the rest as-is (for now at
least).

Finally, start adding some of the pins.