Fix stacking of tooltips in Murder round
Need to set a z-index so they actually float above other things and push
positioning down onto the image elements. That unfortunately costs us
the rotation on the western tooltips, but leaving the transform on the
link/button creates a new stacking context, so we don't get to have nice
things.