Switch to a subsetted emoji font for most of the site
The Noto Color Emoji font is about 8M compressed, and we are only using
a small set of emoji throughout the site. Therefore, add a subsetted
version of the font that only includes the specific emoji we need, and
separately add a global style helper for puzzle pages that *do* need the
wider set of emoji.

In order to not require manually overriding the font on each page, I
left both the subsetted and the full font in the default font-family
specification, which browsers seem to handle gracefully.