The Vault is an interaction, not a round
We don't really want to have a whole round for the endgame, we just want
to have an interaction page for The Vault. But we don't want to expose
the name of the endgame to client-side code, and we don't want to rework
the hunt structure to support interactions that aren't attached to a
round.

So we rename the slug on the round to "endgame", which is sufficiently
innocuous that we can embed it in client-side code, and we special case
it in both the navbar and the all puzzle list.