Don't use Noto Color Emoji on Safari
Safari and Mobile Safari both seem to completely choke on the Noto Color
Emoji font, regardless of how you try to distribute it. Our goal was
making sure that we can render emoji at all (since Windows and Linux
aren't guaranteed to have a good emoji font), but on Apple platforms
that's not a concern, so just use the system emoji font.

Because I'm considering this a workaround for Safari rather than desired
behavior on all Apple browsers, I use a @media selector that narrows
specifically to Safari and mobile Safari.