Live-update witness icon on Missing Diamond puzzle pages
This takes advantage of the main Missing Diamond round dataset to update
individual puzzle pages when a puzzle is solved. It's a little
heavyweight, but should give us better deduping characteristics on both
the client and the server than a more dedicated dataset.

The witness icon now live updates, but more important to me is that we
don't reveal the full-color "solved" icon prematurely.