Add a short-circuit dev only method to skip a virtual interaction
Now that the interaction logic is in place, it makes it harder to step
through the hunt structure, so add a dev-only "skip" button that
effectively picks a random resulting state.

(This works because all of our finalState computations behave when given
the initial state object. That doesn't have to be true in the general
case, but is in our specific cases, so it's good enough for this week.)