DatasetTailer: arm readyPromise after dispatching backfill
Various places would like to ensure that initial tailed events are
delivered before the readyPromise arms, so that readyPromise can be used
to ignore events that happened in the past, so as to focus on events
that are new since startup.  We did not previously provide those
semantics; now we do.

Note that the websocket server, since it keeps a couple of shared
DatasetTailers for use across various subscriptions, may appear
ready immediately, but it will also attempt to deliver any backlog
immediately as well.