Clean up WebRTCClient via useEffect cleanup callback
This happens any time the stream gets disconnected.  Leaking
RTCPeerConnections is...pretty bad for browser performance, especially
if we're creating additional new ones regularly.

Combined with an unfortunate failure mode observed with websocket
failures and repeated start/stops of the gameplay, this avoids a
thundering herd of websocket connections and WebRTC connection attempts.