Checkpoint progress on the stats page
This adds a client-side entrypoint for generating and displaying stats
from the activity log. It fetches and parses the CSV file, then feeds it
to downstream displays. So far that's just the "big graph" (i.e. the
total-solves-over-time).

A few specific notes:
- I need some ordering for the teams on the big graph, if for no other
  reason than stacking order, so we do now have a team ordering (even
  though I will continue to insist that it is not a *ranking*). It is
  composed of (a) finish order then (b) total puzzles solved (if the
  team didn't finish).
- Each team is also assigned a color for rendering purposes. This is
  calculated using the same logic as the tph-site/spoilr-based stats
  page with a fixed random seed.
- The Big Graph hopefully improves on the affordances of the graph
  historically. I suppressed the legend entirely, because no legend with
  200 entries is going to make any sense. I added radio buttons to
  select whether you want to see all puzzles, all metapuzzles, or just
  supermetas. I also added the ability to show/hide teams (all teams are
  shown by default, but it's easy to toggle all-on or all-off) and
  highlighting a subset of teams (which causes other teams to be colored
  in gray)

The performance of this page is not amazing; it takes about 1.5 seconds
on my machine to do the initial render of everything. Most of that time
is being spent parsing the CSV, and specifically parsing the timestamps.
We theoretically could use something lighter weight than Luxon, but I
really don't want to have to deal with timezones in Javascript by hand.