Avoid flash-of-prerendered-content in archive mode
By injecting a script early in the page load which hides the root div
until code is fully loaded, we don't have to worry about flashing the
pre-rendered complete hunt instead of showing the hunt as represented in
localStorage. For user agents that represent search engine crawlers, we
skip this behavior.