Update new-ketchup to work in archive mode
This requires a little bit of shenanigans to generate the case file.

We can't link to a case_file.txt URL, because that won't be dynamically
generated against localStorage log records. (And I'm pretty sure we
can't even have it serve HTML content that uses Javascript to generate
the case file, because the archives will want to actually serve it as
text/plain. Plus nothing will link to it, so wget -R will miss it).

Instead, add archive-only logic to the client that looks for links in
the chat history and replaces them (there will only ever be one) with a
data: URL that contains the contents of the dossier.