Introduce PrefixedLocalStorage to avoid pollution
Since this is eventually going into the puzzles.mit.edu archives, we
want to namespace our localStorage variables so that they don't conflict
with other years' archives.

To do that, introduce a new PrefixedLocalStorage class that takes a
prefix and an underlying storage engine. (The underlying storage engine
is passed as a function so that it doesn't try to resolve in SSR
contexts.). It mostly looks like a Storage implementation (although it
doesn't support accessing values as property members instead of through
getItem - shhh), but it also has its own built-in `addEventListener`
that rewrites events appropriately rather than needing to set a
window-global handler.

Switch all top-level uses of localStorage to use this. And because it's
pretty easy to do so with this abstraction, switch few-nail, lean-lock,
and legitimate-bridge to use this instead of effectively reimplementing
it themselves.