Fix rendering of round page on Background Check
For other rounds where render status changes based on puzzle solve
state, we were already looking at `typeof window` to decide if we were
doing server-side rendering (i.e. hard-code to a fixed state) or
client-side rendering (i.e. match the original hunt behavior), but
Background Check missed that conversion.

Instead of re-implementing similar logic, add a new
`teamIsImmutableForSSR` helper method, which only classifies the public
access team as immutable if (a) we're not in archive mode OR (b) we are
server-side rendering. For client-side rendering or production, treat
all teams as mutable.