Dynamically update round backlinks in archive mode
Puzzles that are released via Stray Leads will get SSR'd with the round
they eventually land in in the round backlink. The styling is a bit
unfortunate, but hard to fix because we'd basically have to re-render
the entire puzzle page on the client. The backlink, though, is
relatively easy to fix.

Add an attribute that lets us identify the round backlink, and then use
a subscription to team_state to update it as needed.